gve: fix XDP cleaning logic when counters overflow
Currently, when XDP cleans completed descriptors, it looks directly at
the counters from the NIC and stored in the tx queue during processing.
The NIC counter should always be equal to or slightly ahead of the tx
queue counter. As such, when the NIC counter wraps, the tx counter will
be much greater than the NIC counter, causing no cleaning to be done,
and in infinite cycle of repolling.

Fixes: 75eaae158b1b ("gve: Add XDP DROP and TX support for GQI-QPL format")
Signed-off-by: Joshua Washington <joshwash@google.com>