gve: keep registry of zc xsk pools in netdev_priv
Relying on xsk_get_pool_from_qid for getting whether zero copy is
enabled on a queue is erroneous, as an XSK pool is registered in
xp_assign_dev whether AF_XDP zero-copy is enabled or not. This becomes
problematic when queues are restarted in copy mode, as all RX queues
with XSKs will register a pool without DMA mapping, which can lead to
kernel panics. This patch adds a bitmap to keep track of which queues
have zero-copy enabled.

Change-Id: If09390b6aebf2e538829f60c169e053acaae373e
Signed-off-by: Joshua Washington <joshwash@google.com>
Google-Bug-Id: 375638380