gve: add XDP_TX and XDP_REDIRECT support for DQ
This patch adds support for XDP_TX and XDP_REDIRECT for the DQ queue
format. To appropriately support transmission of XDP frames, a new
pending packet type GVE_TX_PENDING_PACKET_DQO_XDP_FRAME is introduced
for completion handling, as there was a previous assumption that
completed packets would be SKBs.

XDP_TX handling completes the basic XDP actions, so the feature is
recorded accordingly. This patch also enables the ndo_xdp_xmit callback
allowing DQ to handle XDP_REDIRECT packets originating from another
interface.

The XDP spinlock is moved to common TX ring fields so that it can be
used in both GQ and DQ. Originally, it was in a section which was
mutually exclusive for GQ and DQ.

In summary, 3 XDP features are exposed for the DQ queue format:
1) NETDEV_XDP_ACT_BASIC
2) NETDEV_XDP_ACT_NDO_XMIT
3) NETDEV_XDP_ACT_REDIRECT

Change-Id: Ie06dce241fb6967bfdedf3161c0d474df03587e0
Signed-off-by: Praveen Kaligineedi <pkaligineedi@google.com>
Signed-off-by: Joshua Washington <joshwash@google.com>
Google-Bug-Id: 357627453