gve: share napi for RX and XSK TX
When busy polling is enabled, xsk_sendmsg for AF_XDP zero copy marks
the NAPI ID corresponding to the memory pool allocated for the socket.
In GVE, this NAPI ID will never correspond to the NAPI ID of the
dedicated XDP TX queues registered with the umem because XDP TX is not
set up to share a NAPI with a corresponding RX queue.

This change updates the dedicated XDP TX queues to share the same NAPI
as their corresponding RX queues. The gve_xsk_wakeup callback is updated
to use the RX ring instead of the XDP TX ring to schedule the NAPI
because it simplifies the logic and the NAPI is the same. The branch on
if it is a TX wakeup is removed, as the NAPI poll should be invoked
whether it is a TX wakeup or an RX wakeup.

Note that the notification block will remain separate for the TX
cleanup, keeping that behavior more in line with normal TX queues.

Fixes: fd8e40321a12 ("gve: Add AF_XDP zero-copy support for GQI-QPL format")
Cc: stable@vger.kernel.org
Change-Id: Ib3676a384b4dcf30db1fd41cf053c3bc9269c825
Signed-off-by: Praveen Kaligineedi <pkaligineedi@google.com>
Signed-off-by: Joshua Washington <joshwash@google.com>
Google-Bug-Id: 357627453