gve: merge packet buffer size fields
The data_buffer_size_dqo field in gve_priv and the packet_buffer_size
field in gve_rx_ring theoretically have the same meaning, but they are
defined in two different places and used in two separate contexts. There
is no good reason for this, so this change merges those fields into the
packet_buffer_size field in the RX ring.

This change also introduces a packet_buffer_size field to gve_rx_queue
struct to account for cases where queues are not allocated, such as when
the interface is down.

Change-Id: I2dd0506d9d7753f87460cf6ecbcfdd3bab62e9aa
Signed-off-by: Joshua Washington <joshwash@google.com>