gve: remove skb pointer from gve_tx_fill_pkt_desc_dqo args
Before this patch, gve_tx_fill_pkt_desc_dqo relies on a SKB pointer to
get whether checksum offloading should be enabled. This won't work for
the XDP case, which does not have a SKB. This patch updates the method
to use a boolean representing whether checksum offloading should be
enabled directly.

Change-Id: Ibc7462225805d4f3ea7ab713ce6fa51bbda71084
Signed-off-by: Joshua Washington <joshwash@google.com>
Signed-off-by: Praveen Kaligineedi <pkaligineedi@google.com>
Google-Bug-Id: 357627453